Edited Player prefab so that the rope bug is fixed. Edited Rocket script and prefab so that it works again with collisions (due to issues with module spawn triggers).
RopeSystem:
Changed a "GetMouseButton" for a "GetMouseButtonDown" to prevent unnecessary resource use.
Added a check in Update to prevent unnecessary resource use.
Added a raycast check to HandleInput so that the impulse upwards is only done if touching the ground.
Disabled HandleRopeLength to revent inputs changing rope length.
Disabled HandleRopeUnwrap because the rope currently can't be wrapped.
Reduced redundancies in code.
Reimplemented Ivan's changes.